Extract picker inputs out into shared comp
- started `src/common-ui/GenericPicker` to hold this shared code
- really did not have to change anything with this one, though I did move the loading logic inside the actual SearchInput component
