+++ src/common-ui/GenericPicker/types.ts
@@ -1,0 +1,7 @@
+export type KeyEvent =
+    | 'Enter'
+    | 'ArrowUp'
+    | 'ArrowDown'
+    | ','
+    | 'Tab'
+    | 'Backspace'
